buildguide: add a note describing the possible error with libevdev on FreeBSD
There seem to be a "port" of GNU/Linux's libevdev on FreeBSD.
However, it's not exactly a port, and it includes linux/input.h
file which on FreeBSD doesn't exist. The equivalent file for
linux/input.h is dev/evdev/input.h.

Unless the upstream maintainer of libevdev adds this (may not), we
should add a note here, describing the possible error on compilation
of 86Box on FreeBSD.