network: vde_switch does not require root
Just a bit of erronous documentation.  VDE's whole purpose was to run
a networking stack as a user without requiring root; vde_switch
defaults to setting the UID/GID of the socket to the user that
initiated it with mode 0600: plenty good enough for most users (and
86Box usage).

The default socket directory is already /tmp/vde.ctl, too, but might
as well keep it in the documentation here so that it's more obvious to
change.